Add a simple scheduler for animation frames
This CL adds a simple scheduler for animation frames that holds a given interval
in a given phase.  The scheduler can be told the estimated duration of the
BeginFrame callback and can adjust the callback to hit a given vsync interval.

This CL also makes ganesh_app use this scheduler to animate.

R=jamesr@chromium.org, ernstm@chromium.org

Review URL: https://codereview.chromium.org/731893002