Remove compositing states from DocumentLifeCycle.
We not longer have compositing, so these states are gone.
This was causing asserts to hit because we'd go from
PaintInvalidationClean to InPaintInvalidation, which is
now a valid transition if a new frame is pumped without
style or layout being dirtied.

Unfortunately, we can't test this until we fix test_sky
to actually pump fake frames.

R=esprehn@chromium.org

Review URL: https://codereview.chromium.org/755843003