Compute maxIntrinsicWidth based on the width of styled runs added to the line breaker (#6281)
maxIntrinsicWidth represents the width of the paragraph if no word wrap is
applied.  This can be calculated by summing the width of all runs within each
block of text delimited by a hard line break.  maxIntrinsicWidth is the
maximum width among the blocks in the paragraph.

Fixes https://github.com/flutter/flutter/issues/21965