Add fallback for GetAbsoluteFilePath in UWP (#25379)
In UWP, GetFinalPathNameByHandle requires the app to declare appropriate
capabilities in the app's package manifest. Some of these capabilities are not
permitted in shipping apps on the app store, but may be fine for
development/debugging scenarios. If we can't resolve the full path due to
insufficient access, but have verified the handle is valid, return the original
path.

This gets hot reload requests working in the UWP embedder.

https://github.com/flutter/flutter/issues/79609