Adding DDC module system targets to web SDK artifacts. (#47783)
This is part of an ongoing effort to replace DDC's module system in Flutter Web (for hot reload support).

Side effects of this change include doubling the code generated by DDC whenever this build pathway is hit (until the full module system transition lands), but this shouldn't affect debug performance. We may be able to lock this behind a build flag if this is an issue. 

See related issue: https://github.com/dart-lang/sdk/issues/52361