[Impeller] use smaller SkFont size for determining glyph bounds. (#52868)
Applies the fix that @jason-simmons  previously suggested in https://github.com/flutter/flutter/issues/128624#issuecomment-1846473830 

I'm not sure if these results are better or not ... but I think we try and make this change to see if things are improved anyway? Leaving up golden diffs for others to look at.